pyproject.toml
[tool.poetry]
name = "algorand-contracts-am"
version = "0.1.0"
description = "Smart contracts for AlgoMingle secure messaging platform"
authors = ["fbratten <jack.bratten@adaptivearts.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beaker-pyteal = "^1.1.1"
pyteal = "^0.24.1"
py-algorand-sdk = "^2.5.0"
python-dotenv = "^1.0.0"
algokit-utils = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
